refactor: Rename component folder to components folder and update imports
