+++ src/app/main/apps/network/components/CardItem.js
