+++ src/app/main/apps/network/components/CopyText.js
