+++ src/app/main/apps/network/components/TxItem.js
