+++ src/app/main/apps/network/components/TxLink.js
