+++ src/app/main/apps/network/components/TxStatus.js
