+++ src/app/main/apps/network/components/index.js
